refactor(api): split api functions by domain
