+++ src/lib/api/common.ts
@@ -1,0 +1,2 @@
+export const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:8000';
+export const API_KEY = import.meta.env.VITE_API_KEY;
