+++ src/lib/api/index.ts
@@ -1,0 +1,3 @@
+export * from './items';
+export * from './houses';
+export * from './rooms';
